ちょこっとだけ修正

+++ 寿司チョコレート-流れ魚/寿司チョコレート-流れ魚/ObjBackground.cpp
@@ -10,7 +10,7 @@ using namespace GameL;
 
 float CObjBackground::speed_back()
 {
-	return n;
+	return (float)n;
 }
 
 bool CObjBackground::end_point()
